+++ TextMark/Controllers/Admin_RolesTBController.cs
@@ -8,10 +8,11 @@ using TextMark.Data;
 using TextMark.Models;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Http;
 
 namespace TextMark.Controllers
 {
-    [Authorize]
+    //[Authorize]
     public class Admin_RolesTBController : Controller
     {
         private readonly TextMarkContext _context;
@@ -22,13 +23,44 @@ namespace TextMark.Controllers
 
         public async Task<IActionResult> Index()
         {
+            if (!IsValidUser())
+            {
+                return RedirectToAction("Index", "Home");
+            }
+
             return View(await _context.Roles_TB.ToListAsync());
-            // return View();
+            
+        }
+        
+        private bool IsValidUser()
+        {
+
+            string usertype = "";
+
+            if (HttpContext.Session.GetString("UserType") != null)
+            {
+                usertype = HttpContext.Session.GetString("UserType").ToUpper();
+            }
+
+
+            if (usertype != "ADMIN")
+            {
+                return false;
+            }
+            else
+            {
+                return true;
+            }
+
         }
         // GET: Logins/Create
         public IActionResult Create()
         {
-         //   Select_All_Roles();
+            if (!IsValidUser())
+            {
+                return RedirectToAction("Index", "Home");
+            }
+            //   Select_All_Roles();
             return View();
         }
 
@@ -39,6 +71,11 @@ namespace TextMark.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("Role_ID,Role_Text")] Roles_TB roles_tb)
         {
+            if (!IsValidUser())
+            {
+                return RedirectToAction("Index", "Home");
+            }
+
             if (ModelState.IsValid)
             {
                 _context.Add(roles_tb);
@@ -55,6 +92,11 @@ namespace TextMark.Controllers
         // GET: Logins/Details/5
         public async Task<IActionResult> Details(int? id)
         {
+            if (!IsValidUser())
+            {
+                return RedirectToAction("Index", "Home");
+            }
+
             if (id == null)
             {
                 return NotFound();
@@ -72,7 +114,12 @@ namespace TextMark.Controllers
         // GET: Logins/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
-           // Select_All_Roles();
+            if (!IsValidUser())
+            {
+                return RedirectToAction("Index", "Home");
+            }
+
+            // Select_All_Roles();
             if (id == null)
             {
                 return NotFound();
@@ -93,6 +140,11 @@ namespace TextMark.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("Role_ID,Role_Text")] Roles_TB Roles_tb)
         {
+            if (!IsValidUser())
+            {
+                return RedirectToAction("Index", "Home");
+            }
+
             if (id != Roles_tb.Role_ID)
             {
                 return NotFound();
@@ -129,6 +181,11 @@ namespace TextMark.Controllers
         // GET: Logins/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
+            if (!IsValidUser())
+            {
+                return RedirectToAction("Index", "Home");
+            }
+
             if (id == null)
             {
                 return NotFound();
@@ -149,6 +206,11 @@ namespace TextMark.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            if (!IsValidUser())
+            {
+                return RedirectToAction("Index", "Home");
+            }
+
             var Role = await _context.Roles_TB.FindAsync(id);
             _context.Roles_TB.Remove(Role);
             await _context.SaveChangesAsync();
